refactor: 구매하기 버튼 이벤트를 추가하는 부분을 PurchaseEvent.js로 분리
